# shared-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Property Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/contracts": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Create a new rental contract",
        "description": "Creates an active contract for a room+tenant.  Enforces BR-01 (one active per room) and BR-02 (min deposit).",
        "operationId": "create_contract_api_v1_contracts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/contracts/active": {
      "get": {
        "tags": [
          "contracts"
        ],
        "summary": "List active contracts",
        "description": "Returns all active contracts, optionally filtered by property_id.",
        "operationId": "list_active_contracts_api_v1_contracts_active_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by property",
              "title": "Property Id"
            },
            "description": "Filter by property"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}": {
      "get": {
        "tags": [
          "contracts"
        ],
        "summary": "Get a contract by ID",
        "description": "Returns full contract details including termination and extensions.",
        "operationId": "get_contract_api_v1_contracts__contract_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}/terminate": {
      "patch": {
        "tags": [
          "contracts"
        ],
        "summary": "Terminate an active contract",
        "description": "Terminates an active contract (BR-04).  Requires reason.  Sets room to available.",
        "operationId": "terminate_contract_api_v1_contracts__contract_id__terminate_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminateContractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}/extend": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Extend a contract's lease",
        "description": "Extends the end date of an active contract and records the extension.",
        "operationId": "extend_lease_api_v1_contracts__contract_id__extend_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendLeaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}/renew": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Renew a terminated/expired contract",
        "description": "Creates a new contract from a terminated/expired original.  Same room/tenant, updated terms.",
        "operationId": "renew_contract_api_v1_contracts__contract_id__renew_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewContractRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leases/{room_id}/history": {
      "get": {
        "tags": [
          "contracts"
        ],
        "summary": "Get lease history for a room",
        "description": "Returns all contracts (past and present) for a given room, newest first.",
        "operationId": "get_lease_history_api_v1_leases__room_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaseHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContractCreateResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ContractResponse"
          },
          "meta": {
            "type": "null",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ContractCreateResponse",
        "description": "Wrapper for 201 CREATED response (SDD \u00a73.1)."
      },
      "ContractListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ContractResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ContractListResponse",
        "description": "Paginated list response for contracts (SDD \u00a73.1)."
      },
      "ContractResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "room_id": {
            "type": "string",
            "format": "uuid",
            "title": "Room Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "monthly_rent": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Monthly Rent"
          },
          "deposit_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Deposit Amount"
          },
          "status": {
            "$ref": "#/components/schemas/ContractStatus"
          },
          "special_conditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Conditions"
          },
          "is_renewal": {
            "type": "boolean",
            "title": "Is Renewal",
            "default": false
          },
          "renewed_from_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Renewed From Id"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "termination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractTerminationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "extensions": {
            "items": {
              "$ref": "#/components/schemas/LeaseExtensionResponse"
            },
            "type": "array",
            "title": "Extensions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "room_id",
          "tenant_id",
          "property_id",
          "start_date",
          "end_date",
          "monthly_rent",
          "deposit_amount",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ContractResponse",
        "description": "Response body for a single Contract resource (SDD \u00a73.1 format)."
      },
      "ContractStatus": {
        "type": "string",
        "enum": [
          "active",
          "terminated",
          "expired",
          "draft"
        ],
        "title": "ContractStatus",
        "description": "Possible states for a Contract (SDD \u00a76.2: State Machines).\n\nTransitions:\n    draft \u2500\u2500\u2192 active (via signing)\n    active \u2500\u2500\u2192 terminated (via terminate_contract, BR-04)\n    active \u2500\u2500\u2192 expired (auto via scheduler when end_date < today)\n    terminated \u2500\u2500\u2192 (immutable, new contract created for renew)\n    expired \u2500\u2500\u2192 (immutable)"
      },
      "ContractTerminationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "termination_date": {
            "type": "string",
            "format": "date",
            "title": "Termination Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "terminated_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminated By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reason",
          "termination_date",
          "created_at"
        ],
        "title": "ContractTerminationResponse",
        "description": "Termination record embedded in ContractResponse (SDD \u00a75.2)."
      },
      "CreateContractRequest": {
        "properties": {
          "room_id": {
            "type": "string",
            "format": "uuid",
            "title": "Room Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "monthly_rent": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Monthly Rent"
          },
          "deposit_amount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Deposit Amount"
          },
          "special_conditions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Conditions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "room_id",
          "tenant_id",
          "property_id",
          "start_date",
          "end_date",
          "monthly_rent",
          "deposit_amount"
        ],
        "title": "CreateContractRequest",
        "description": "Request body for POST /api/v1/contracts (SDD \u00a73.3).\n\nValidation:\n    - ``end_date`` must be after ``start_date``.\n    - ``monthly_rent`` must be positive.\n    - ``deposit_amount`` must be >= 0.\n    - ``start_date`` must not be in the past (service layer also checks).\n    - BR-01/BR-02 enforced in service layer (not here)."
      },
...
      "ExtendLeaseRequest": {
        "properties": {
          "new_end_date": {
            "type": "string",
            "format": "date",
            "title": "New End Date"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "new_end_date"
        ],
        "title": "ExtendLeaseRequest",
        "description": "Request body for POST /api/v1/contracts/{id}/extend.\n\nValidation:\n    - ``new_end_date`` must be after the current end_date."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeaseExtensionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "previous_end_date": {
            "type": "string",
            "format": "date",
            "title": "Previous End Date"
          },
          "extended_to": {
            "type": "string",
            "format": "date",
            "title": "Extended To"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "extended_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extended By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "previous_end_date",
          "extended_to",
          "created_at"
        ],
        "title": "LeaseExtensionResponse",
        "description": "Lease extension record embedded in ContractResponse."
      },
      "LeaseHistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "monthly_rent": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Monthly Rent"
          },
          "status": {
            "$ref": "#/components/schemas/ContractStatus"
          },
          "is_renewal": {
            "type": "boolean",
            "title": "Is Renewal"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "termination_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termination Reason"
          },
          "termination_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termination Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "start_date",
          "end_date",
          "monthly_rent",
          "status",
          "is_renewal",
          "created_at"
        ],
        "title": "LeaseHistoryItem",
        "description": "A single lease history entry for a room."
      },
      "LeaseHistoryResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/LeaseHistoryItem"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "type": "null",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "LeaseHistoryResponse",
        "description": "Response for GET /api/v1/leases/{room_id}/history."
      },
...
      "RenewContractRequest": {
        "properties": {
          "new_start_date": {
            "type": "string",
            "format": "date",
            "title": "New Start Date"
          },
          "new_end_date": {
            "type": "string",
            "format": "date",
            "title": "New End Date"
          },
          "new_monthly_rent": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "New Monthly Rent"
          },
          "new_deposit_amount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "New Deposit Amount"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "new_start_date",
          "new_end_date",
          "new_monthly_rent",
          "new_deposit_amount"
        ],
        "title": "RenewContractRequest",
        "description": "Request body for POST /api/v1/contracts/{id}/renew.\n\nCreates a new contract record based on the original.  The original\ncontract must be in ``terminated`` or ``expired`` state (CONT-005)."
      },
...
      "TerminateContractRequest": {
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/TerminationReason"
          },
          "termination_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termination Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "TerminateContractRequest",
        "description": "Request body for POST /api/v1/contracts/{id}/terminate (SDD \u00a73.3).\n\nValidation:\n    - ``reason`` is required and must be non-empty.\n    - ``termination_date`` defaults to today if not provided."
      },
      "TerminationReason": {
        "type": "string",
        "enum": [
          "tenant_moved_out",
          "owner_terminated",
          "breach_of_contract",
          "mutual_agreement",
          "other"
        ],
        "title": "TerminationReason",
        "description": "Standardised reasons for contract termination (SDD \u00a75.2)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}